# docs/swagger.yaml
basePath: /api
definitions:
  dtos.AcceptInviteRequest:
    properties:
      token:
        example: abc123def456...
        type: string
    required:
    - token
    type: object
...
  dtos.ChangeMemberRoleRequest:
    properties:
      role:
        enum:
        - admin
        - member
        example: admin
        type: string
    required:
    - role
    type: object
...
  dtos.CreateOrganizationRequest:
    properties:
      name:
        example: Acme Corp
        maxLength: 80
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dtos.InviteToOrganizationRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
...
  dtos.UpdateOrganizationNameRequest:
    properties:
      name:
        example: New Acme Corp
        maxLength: 80
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  errors.BasicError:
    properties:
      error:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.CurrentUserOrganizationsResponse:
    properties:
      id:
        description: holds org id and not orgUserID
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  models.Organization:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        maxLength: 100
        minLength: 4
        type: string
      owner_id:
        type: string
    required:
    - name
    - owner_id
    type: object
  models.OrganizationInvite:
    properties:
      createdAt:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      invited_by:
        type: string
      org_id:
        type: string
      status:
        enum:
        - invited
        - accepted
        - revoked
        type: string
      token:
        type: string
      updatedAt:
        type: string
    required:
    - email
    - expires_at
    - invited_by
    - status
    - token
    type: object
  models.OrganizationUser:
    properties:
      id:
        type: string
      joined_at:
        type: string
      org_id:
        type: string
      role:
        enum:
        - owner
        - admin
        - member
        type: string
      user_id:
        type: string
    required:
    - joined_at
    - org_id
    - role
    - user_id
    type: object
  models.OrganizationUserResult:
    properties:
      image_url:
        type: string
      joined_at:
        type: string
      name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  models.OrganizationWithMembersCount:
    properties:
      created_at:
        type: string
      id:
        type: string
      members_count:
        type: integer
      name:
        type: string
      owner_id:
        type: string
    type: object
...
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  utils.SuccessResponseWithTotal:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      total:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@taskio.com
    name: API Support
  description: Task management system API with real-time features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taskio API
  version: "1.0"
paths:
...
  /orgs:
    get:
      consumes:
      - application/json
      description: Retrieves all organizations where the user is a member
      produces:
      - application/json
      responses:
        "200":
          description: Organizations fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CurrentUserOrganizationsResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get user's organizations
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      description: Creates a new organization (user becomes owner)
      parameters:
      - description: Organization details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Organization'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - Organizations
  /orgs/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an organization and all related data (owner
        only)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Only owner can delete
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organizations
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific organization
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OrganizationWithMembersCount'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - Organizations
    patch:
      consumes:
      - application/json
      description: Updates the name of an organization (owner or admin only)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New organization name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateOrganizationNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Update organization name
      tags:
      - Organizations
  /orgs/{id}/invite:
    post:
      consumes:
      - application/json
      description: Sends an email invitation to join the organization (owner or admin
        only)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Email to invite
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.InviteToOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation sent successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Invite user to organization
      tags:
      - Organizations
  /orgs/{id}/members:
    get:
      consumes:
      - application/json
      description: Retrieves all members of a specific organization
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Members fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.OrganizationUserResult'
                  type: array
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get organization members
      tags:
      - Organizations
...
  /orgs/{id}/users/{orgUserId}/role:
    put:
      consumes:
      - application/json
      description: Changes the role of an organization member (owner or admin only)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Organization User ID (UUID)
        in: path
        name: orgUserId
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ChangeMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OrganizationUser'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - Organizations
  /orgs/invites/accept:
    post:
      consumes:
      - application/json
      description: Accepts a pending invitation to join an organization
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.AcceptInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OrganizationUser'
              type: object
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Accept organization invitation
      tags:
      - Organizations
  /orgs/invites/pending:
    get:
      consumes:
      - application/json
      description: Retrieves all pending organization invitations for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Invitations fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.OrganizationInvite'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get pending invitations
      tags:
      - Organizations
  /orgs/invites/reject:
    post:
      consumes:
      - application/json
      description: Rejects a pending invitation to join an organization
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.AcceptInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation rejected
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Reject organization invitation
      tags:
      - Organizations
...
securityDefinitions:
  BearerAuth:
    description: JWT token stored in cookie
    in: cookie
    name: taskio-cookie
    type: apiKey
swagger: "2.0"